editorder added, emptying shopping cart after order placed

+++ OnlineStore/View/Seller/ProductManagement.swift
@@ -13,19 +13,24 @@ struct ProductManagement: View {
     @FetchRequest(sortDescriptors: []) private var products: FetchedResults<Product>
     
     var body: some View {
-        List {
-            ForEach(productVM.products) { product in
-                NavigationLink(destination: EditProduct(product: product)) {
-                    Text(product.title ?? "Unknown Title")
+        if productVM.products.isEmpty {
+            Text("There are currently no products")
+                .navigationTitle("Edit Product")
+        } else {
+            List {
+                ForEach(productVM.products) { product in
+                    NavigationLink(destination: EditProduct(product: product)) {
+                        Text(product.title ?? "Unknown Title")
+                    }
                 }
+                .onDelete(perform: { indexSet in
+                    productVM.removeProduct(at: indexSet, products: products, viewContext: viewContext)
+                })
             }
-            .onDelete(perform: { indexSet in
-                productVM.removeProduct(at: indexSet, products: products, viewContext: viewContext)
-            })
-        }
-        .navigationTitle("Edit Product")
-        .toolbar {
-            EditButton()
+            .navigationTitle("Edit Product")
+            .toolbar {
+                EditButton()
+            }
         }
     }
 }
